refactor: search input value hooks로 분리
