+++ src/components/common/SearchInput/index.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import { Search, X } from 'lucide-react';
+import { Search } from 'lucide-react';
 import { BaseInput } from '../BaseInput';
 import type { SearchInputProps } from '@/types/input';
-import { useState, useCallback } from 'react';
-import { useDebouncedCallback } from 'use-debounce';
 import type React from 'react';
+import Image from 'next/image';
+import { useSearchValue } from '@/hooks/useSearchValue';
 
 export function SearchInput({
   onSearch,
@@ -14,32 +14,30 @@ export function SearchInput({
   loading,
   ...props
 }: SearchInputProps) {
-  const [value, setValue] = useState('');
-
-  const debouncedSearch = useDebouncedCallback((value: string) => {
-    onSearch?.(value);
-  }, 500);
-
-  const handleChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      const newValue = e.target.value;
-      setValue(newValue);
-      debouncedSearch(newValue);
-    },
-    [debouncedSearch],
-  );
+  const { value, setValue, handleChange } = useSearchValue(onSearch);
 
   return (
     <div className="relative">
       <BaseInput
         endIcon={
-          value ? <X className="h-4 w-4" /> : <Search className="h-6 w-6" />
+          value ? (
+            <Image
+              src="/images/svg/delete.svg"
+              width={16}
+              height={16}
+              alt="search value delete"
+              onClick={() => setValue('')}
+              className="cursor-pointer"
+            />
+          ) : (
+            <Search className="h-6 w-6" />
+          )
         }
         value={value}
         placeholder={placeholder}
         onChange={handleChange}
         onClick={onClick}
-        className="w-[330px] h-[55px] bg-white rounded-[10px] border-[#FFA914] border-[1px]"
+        className="h-[55px] bg-white rounded-[10px] border-[#FFA914] border-[1px]"
         {...props}
       />
       {loading && (
